Update VLM.py
Updated VLM to account for non-planar wings. Uses dot product between the normal vector of the wake trace and the induced velocity vector to calculate the normal velocity to the wake trace. Also now integrates along the wake trace using variable "s" which is the distance between control points (not just spanwise distance between control points as was used earlier).